Release lock during task scheduling, tweak token mechanism
Releasing the lock avoids blocking various Execution instance
methods when dealing with a synchronous task scheduler. If an
execution is frozen while tasks are in the process of being
scheduled, we reset the nodes (in the execution snapshot) to READY.

Loosens the ScheduledTaskToken validity requirements to make
life easier for schedulers.